Move src code into page, hooks and components
