+++ src/components/Alignment/RandomAlignmentTable.tsx
