+++ src/components/Alignment/RandomAlignmentTables.tsx
