+++ src/components/Alignment/alignmentRelations.ts
