+++ src/components/Formatting/toInitials.test.ts
