+++ src/components/Formatting/toInitials.ts
