+++ src/components/Stats/HeroStatsGenerator.tsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
 
-import { heroStatsGenerator } from "../pages/api/dice/hero-stats-generator";
+import { heroStatsGenerator } from "../../pages/api/dice/hero-stats-generator";
 
 import { Table } from "../Layout/Table";
