+++ src/pages/404.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import Head from "next/head";
 
-import { Explanation } from "../Layout/Explanation";
-import { Main } from "../Layout/Main";
+import { Explanation } from "../components/Layout/Explanation";
+import { Main } from "../components/Layout/Main";
 
 const Page = () => (
   <>
